requirements.txt
asyncio==3.4.3
discord.py==2.1.0
#Pillow==9.0.0
psycopg2-binary==2.8.1
#PyNaCl==1.3.0
requests==2.25.1
urllib3==1.26.5
latex==0.7.0
deepmerge==0.1.0
#youtube-dl==2019.5.11
